requirements.txt
openai==2.32.0
atproto==0.0.65
httpx==0.28.1
ddgs==9.14.1
pillow==11.3.0
matplotlib==3.10.0
python-dotenv==1.0.1
streamlit==1.41.1
scikit-learn==1.6.1

dnspython==2.8.0
libipld==3.3.2
primp==1.2.3
lxml==6.1.0